+++ test/runtests.jl
@@ -8,5 +8,5 @@ include("test_selection.jl")
 include("test_mutation.jl")
 include("test_fitness.jl")
 include("test_genetic_algorithms.jl")
-include("test_utils.jl")
+include("test_visualization.jl")
